Extract interface NftApi
