+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/nft/INftApiProvider.kt
@@ -1,0 +1,9 @@
+package io.horizontalsystems.bankwallet.modules.nft
+
+import io.horizontalsystems.bankwallet.entities.Account
+import io.horizontalsystems.bankwallet.entities.Address
+
+interface INftApiProvider {
+    suspend fun getCollectionRecords(address: Address, account: Account): List<NftCollectionRecord>
+    suspend fun getAssetRecords(address: Address, account: Account): List<NftAssetRecord>
+}
